+++ src/utils/config.ts
@@ -2,10 +2,15 @@ import { SdkBaseChainConfigParams } from '@connext/nxtp-sdk'
 import { NetworkMetadata, NetworkMap, TokenMetadataMap } from '@/config/types'
 import AVAXIcon from '@/assets/token-logos/AVAX.png'
 
-export const config = await import('@nomad-xyz/sdk-bridge').then(async({ BridgeContext }) => {
-  const { conf } = (await BridgeContext.fetch(process.env.VUE_APP_NOMAD_ENVIRONMENT, false))
-  return conf
-})
+export const config = await import('@nomad-xyz/sdk-bridge').then(
+  async ({ BridgeContext }) => {
+    const { conf } = await BridgeContext.fetch(
+      process.env.VUE_APP_NOMAD_ENVIRONMENT,
+      false
+    )
+    return conf
+  }
+)
 
 const {
   VUE_APP_ETHEREUM_RPC,
@@ -66,9 +71,7 @@ export const getConnextConfig = (
   return connextConfig
 }
 
-export const getNetworksConfig = (
-  tokens: TokenMetadataMap
-): NetworkMap => {
+export const getNetworksConfig = (tokens: TokenMetadataMap): NetworkMap => {
   const networks: NetworkMap = {}
 
   Object.keys(config.bridgeGui).forEach((networkName) => {